Feat: let the user order products
There are many changes to accomplish this feature.

First, the router.jsx has dynamic segments support. This was changed to
ease the navigation between pages.

The app.jsx is the most changed file. First, it has a custom hook to
fetch products data. Placing it here and passing it to Shop component as
a prop prevents to fetch data each time the user enters the shop page.

Another important feat is state variables. These are self explanatory,
it keeps data between renders and like the custom hook, keeping them in
app.jsx prevents that the values are reset.

Related with the router.jsx changes, this file now has conditional
rendering. With this, the Navbar is only placed here and there is no
need to render in other components. It also passes props to Shop to
render the products (more of this below).

Speaking of Navbar, this file had a prop called isShop. With the yet
again changed to router.jsx it now needs another variable to keep
allowing the user visiting the shop page and display a checkout button.

The Shop component also receives new props. Some of these props are used
to render the products and let the user to order. For better
"readability", this file now has PropTypes to better know the expected
props and its types, like in other files.

There is a new Card component. This renders each product. At the moment
it has an image, the product title, 2 buttons to adjust the quantity.
Also there is a field to write the desired quantity and a button to
send said quantity to the cart (e.g.the checkout button that displays
the product count).

Lastly, but not least important, changed the README.md file.