Refactor: code in component files and App.jsx
Made several changes in some component files.

The Card component now receives the handleSubmit function in order to
make tests easier. Another change that eases tests is adding a test id
to the div that contains all the elements of each Card. Lastly, in the
handleChange function, there is a condition that adds a 0 to the input
element when the user tries to enter a negative value; this is done to
prevent unexpected behaviours when the user submits those values.

The Checkout component now has some paragraph elements to display a bit
more information about the page.

The Navbar component now receives a function that redirects to the
/checkout page when it's called. As a requirement from the project page,
the Navbar now has the prouct count besides the links when the user is
on the /shop page. The button that redirects the user to /checkout page
no longer possess the product count.

The Shop component has a new function to update the user's product count
each time it clicks a "submit order" button. A minor change is removing
the productCount prop from this component. Now this prop is passed from
the App component to the Navbar component.

The App component now imports a custom hook that was the previous
useEffect used to fetch the products shown on the /shop page. This
change makes tests related with API calls easier.