Add: tests and Mock Service Worker
MSw is used to mock API calls in some tests to return some values or
responses instead of actually making API calls. There are several new
files and some changes in main.jsx to add support of this library.

The test files try to cover the majority of the code spread in all the
components. Many files just contain tests that check if a component
renders certain elements. The best tests are those that cover expected
behaviours from the user when utilizing the page.

In App.test.jsx there is a lack of these kind of tests. At the moment
of writing this, it was hard to mock or render the App component in
order to test these behaviours. Maybe it's needed to refactor the
components or something else.